requirements.txt
sounddevice
numpy
aubio-ledfx
mido
python-rtmidi
scipy
PyYAML
fastapi
uvicorn[standard]
